Выполнено тз 15 спринта

+++ src/main/java/ru/yandex/practicum/shareit/user/UserServiceImpl.java
@@ -2,42 +2,53 @@ package ru.yandex.practicum.shareit.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.yandex.practicum.shareit.exception.ConflictException;
 import ru.yandex.practicum.shareit.exception.NotFoundException;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class UserServiceImpl implements UserService {
 
-    private final UserRepository userRepository;
+    private final UserJpaRepository userRepository;
 
     @Override
     public UserDto create(UserDto user) {
-        if (!userRepository.isUniqueEmail(user.getEmail(), null)) {
+        if (userRepository.existsByEmail(user.getEmail())) {
             throw new ConflictException("Пользователь с такой почтой уже существует!");
         }
-        return UserMapper.toUserDto(userRepository.create(UserMapper.toUser(user)));
+        return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(user)));
     }
 
     @Override
     public UserDto update(UserDto userDto, Long userId) {
-        if (userDto.getEmail() != null &&
-                !userRepository.isUniqueEmail(userDto.getEmail(), userId)) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("Такого пользователя нет!"));
+
+        if (userDto.getEmail() != null
+                && userRepository.existsByEmailAndIdNot(userDto.getEmail(), userId)) {
             throw new ConflictException("Пользователь с такой почтой уже существует!");
         }
-        User user = userRepository.findById(userId);
+
         user.setName(userDto.getName() != null ? userDto.getName() : user.getName());
         user.setEmail(userDto.getEmail() != null ? userDto.getEmail() : user.getEmail());
-        return UserMapper.toUserDto(userRepository.update(user));
+        return UserMapper.toUserDto(userRepository.save(user));
     }
 
     @Override
+    @Transactional(readOnly = true)
     public UserDto findById(Long userId) {
-        return UserMapper.toUserDto(userRepository.findById(userId));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("Такого пользователя нет!"));
+        return UserMapper.toUserDto(user);
     }
 
     @Override
     public void delete(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("Такого пользователя нет!");
+        }
         userRepository.deleteById(userId);
     }
 }
